+Added some basic shader support in the dynamic objects templates:   Current shaders supported:   LIGHTMAP   LIGHTMAP_LIGHTED   TERRAIN   WATER
Terrain and water are custom shaders (GLSL) that can apply the terrain and water(special version) of the terrain to any other meshes.
LIGHTMAP and LIGHTMAP_LIGHTED are based on the default Irrlicht material types. More to come.

Once theses shader are defined in the Template XML they can be displayed with dynamic objects and custom tiles.xml
+ Added a terrain example that use the terrain and water shader.  

git-svn-id: http://svn.code.sf.net/p/irrrpgbuilder/code@477 528fbfa9-60f6-4f4a-a82b-d133dc65ca33